+++ CHANGELOG.md
@@ -1,15 +1,16 @@
 HEAD
 =====
-* Fix bug that caused CMUser's userId field to not be serialized when cached to the filesystem.
-* Fixed NSNull being set in CMUser profile as NSNULL, and not being set as nil. Changed to set as nil.
 
 
 v1.3 (December 11, 2012)
+=====
 * Added Social Login. Login via social networks that are supported through CloudMine and Singly.
   * https://cloudmine.me/docs/social
 * Allow developers to specify scope for the social login.
 * Linking of accounts! Logging in a user through a social network while a user is previously logged in will link the two accounts.
 * Creation of accounts through social login! Logging in through a social network will automatically create the user account if no user is logged in.
+* Fix bug that caused CMUser's userId field to not be serialized when cached to the filesystem.
+* Fixed NSNull being set in CMUser profile as NSNULL, and not being set as nil. Changed to set as nil.
 
 v1.2.2 (November 29, 2012)
 =====