+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 HEAD
 =====
-* Added call to register device push notification token with cloudmine
+* Added call to register device push notification token with cloudmineb
+* Added CMAppDelegate base which can be used to handle push notification registration for you.
 
 v1.3 (December 11, 2012)
 =====